+++ packages/frontend-components/src/components/ProjectList/ProjectPlaceholder.tsx
@@ -11,6 +11,7 @@ const ProjectPlaceholder = (props: CardProps) => (
       </Placeholder.Header>
       <Placeholder.Paragraph>
         <Placeholder.Line />
+        <Placeholder.Line />
       </Placeholder.Paragraph>
     </Placeholder>
   </ProjectPlaceholderWrapper>